% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{draw_boxplot}
\alias{draw_boxplot}
\title{Boxplot of gene from expression file, by group}
\usage{
draw_boxplot(df, gene, grouping)
}
\arguments{
\item{df}{data frame of expression matrix with genes as rownames}

\item{gene}{gene of interest}

\item{grouping}{groups for condition, can be as a factor}
}
\description{
Boxplot of gene from expression file, by group
}
