% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{draw_pca}
\alias{draw_pca}
\title{Draw PCA plot from data frame}
\usage{
draw_pca(
  df,
  log = T,
  grouping = NULL,
  x_axis = "PC1",
  y_axis = "PC2",
  return_data = T
)
}
\arguments{
\item{df}{data frame with samples in columns}

\item{log}{should the data be log2 transformed?}

\item{grouping}{groups present in the samples supplied as character vector}

\item{x_axis}{principle components to be used on the x-axis}

\item{y_axis}{principle components to be used on the y-axis}

\item{return_data}{should the PCA data frame be returned}
}
\description{
Draw PCA plot from data frame
}
\examples{
\dontrun{
draw_pca(expression_matrix, grouping = groups, return_data = F)
}

}
