% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{draw_volcano}
\alias{draw_volcano}
\title{Draw a volcano plot with FDR and logFC values}
\usage{
draw_volcano(data_frame, fdr_cutoff = 0.05, logfc_cutoff = 1)
}
\arguments{
\item{data_frame}{Data frame with logFC, FDR, and gene columns at minimum}

\item{fdr_cutoff}{FDR cutoff for significant points}

\item{logfc_cutoff}{LogFC cutoff for significant points}
}
\description{
Draw a volcano plot with FDR and logFC values
}
\examples{
\dontrun{
draw_volcano(degs, fdr_cutoff = 0.01, logfc_cutoff = 0.7)
}

}
